feat(threaded): support multi-threaded training
A new attribute has been added to the `Trainer` specifying the number
of threads to use during training.

The weight updating/error rate behavior has changed as instead of
updating the weights/error after each example, the weight updates/error
are held until all examples have been processed. At the end of the
training epoch, the weight updates are performed and the error is
updated.